+++ CHANGES.md
@@ -2,3 +2,4 @@
 
 * #150: Reduce the number of threads configured
 * #141: Update commons-io
+* #151: Fixed bug related to query param matching